\name{To-do list}
\Rdversion{1.1}
\alias{todo}
\title{
MRPP Package To-Do List
}
\description{
To-do's and to-try's
}
\details{

\enumerate{
	\item re-implement interfaces using mrpp class
	\item add common importance measure core and interface
	\item \code{\link[Rmpfr:Rmpfr]{Rmpfr}}-based weighting? 
	\item MRBP
	\item allowing some NA's in the distance matrix
	\item double checking one-way design for all variable selection functions
	\item check number of RHS terms in formula interface
	\item allow parallel running
	\item saving the seed in choosing permutations
	\item adding test code for grad.smoothp
}
}
\author{Long Qu}
\keyword{package}
